requirements.txt
# Core deep learning (GPU support)
torch>=2.0.0
torchvision>=0.15.0

# Image & video processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

streamlit>=1.30.0
gradio>=4.0.0

# Utilities
tqdm>=4.66.0
